+++ app/controllers/drivers_controller.rb
@@ -4,6 +4,8 @@ class DriversController < UsersController
   rescue_from ActiveRecord::InvalidForeignKey, with: :invalid_foreign_key
 
   def index
+    return if cannot?(:read, Driver) && unauthorized_redirect
+
     @q = Driver.includes(:enterprise)
                .includes(:person)
                .accessible_by(current_ability)
@@ -13,15 +15,21 @@ class DriversController < UsersController
     @drivers = @q.result(distinct: false)
   end
 
+  def show
+    return if cannot?(:read, Driver) && unauthorized_redirect
+  end
+
   def new
+    return if cannot?(:create, Driver) && unauthorized_redirect
+
     @driver = Driver.new
     @driver.build_person
     @driver.person.build_address
   end
 
-  def show; end
-
   def create
+    return if cannot?(:create, Driver) && unauthorized_redirect
+
     @driver = Driver.new(params_driver)
     @driver.validate_all = true
     @driver.person.validate_all = true
@@ -35,9 +43,13 @@ class DriversController < UsersController
     end
   end
 
-  def edit; end
+  def edit
+    return if cannot?(:update, Driver) && unauthorized_redirect
+  end
 
   def update
+    return if cannot?(:update, Driver) && unauthorized_redirect
+
     @driver.validate_all = true
     @driver.person.validate_all = true
     @driver.person.address.validate_address = true
@@ -51,6 +63,8 @@ class DriversController < UsersController
   end
 
   def destroy
+    return if cannot?(:destroy, Driver) && unauthorized_redirect
+
     bank_accounts = BankAccount.where(person: @driver.person)
 
     if @driver.destroy && bank_accounts.destroy_all && @driver.person.destroy
@@ -62,6 +76,11 @@ class DriversController < UsersController
   end
 
   private
+
+  def unauthorized_redirect
+    redirect_to(root_path)
+    flash[:danger] = 'Você não possui permissão para realizar esta ação.'
+  end
 
   def invalid_foreign_key
     redirect_to(drivers_path)
@@ -97,7 +116,7 @@ class DriversController < UsersController
                   :cnh_type,
                   :is_employee,
                   :is_blocked,
-                  person_attributes: [User::Person.permitted_attributes,
+                  person_attributes: [Person.permitted_attributes,
                                       address_attributes: Address.permitted_attributes,
                                       bank_accounts_attributes: BankAccount.permitted_attributes])
   end
